Java programs better modified and revisited
This contains the java programs for lab experiments 4,5,8,9,16,17.
The already existing code for inheritance (experiment -4) is not done by array-of-objects method and hence there is a chance that the array of objects method implementation will be required. And also, the existing code for abstraction (exp-5) contains only number of sides and does not include method to find area; in this code, that provision is also added.